Refactor: App.tsx Login 함수 부분 리팩토리
