+++ frontend/src/pages/calendar/index.tsx
@@ -1,4 +1,4 @@
-import React, { useState } from "react";
+import React, { useState, useCallback } from "react";
 import FullCalendar from "@fullcalendar/react";
 import dayGridPlugin from "@fullcalendar/daygrid";
 import timeGridPlugin from "@fullcalendar/timegrid";
@@ -10,15 +10,16 @@ import "../../styles/FullCalendar.css";
 const FILTERS = ["전체", "공모전", "자격증", "코딩테스트", "개인"];
 
 const CalendarPage: React.FC = () => {
-  const [selected, setSelected] = useState<string>("전체");
+  const [selected, setSelected] = useState("전체");
 
-  const handleAddSchedule = () => {
+  const handleAddSchedule = useCallback(() => {
     alert("일정 추가 폼/모달이 열립니다!");
-    // 실제로는 모달 상태를 true로 바꾸고, 모달 컴포넌트 띄우기
-  };
+    // TODO: 일정 추가 모달 상태 관리
+  }, []);
 
   return (
     <div className="flex max-w-6xl mx-auto p-6 gap-6 select-none">
+      {/* 좌측: 필터/일정추가 */}
       <aside className="w-52">
         <div className="bg-white rounded-lg shadow p-4 flex flex-col gap-4">
           <AddScheduleButton onClick={handleAddSchedule} />
@@ -29,7 +30,8 @@ const CalendarPage: React.FC = () => {
           />
         </div>
       </aside>
-      <section className="flex-1 min-w-0 bg-white rounded-lg shadow p-4 select-none">
+      {/* 우측: 캘린더 */}
+      <section className="flex-1 min-w-0 bg-white rounded-lg shadow p-4">
         <FullCalendar
           plugins={[dayGridPlugin, timeGridPlugin]}
           initialView="dayGridMonth"
